funcao de excluir galpao ou motoqueiro pronta

+++ src/main/java/challenge_mottu_2_semestre/challenge_mottu/service/GalpaoService.java
@@ -3,6 +3,7 @@ package challenge_mottu_2_semestre.challenge_mottu.service;
 import challenge_mottu_2_semestre.challenge_mottu.model.DTO.GalpaoDTO;
 import challenge_mottu_2_semestre.challenge_mottu.model.Galpao;
 import challenge_mottu_2_semestre.challenge_mottu.repository.GalpaoRepository;
+import challenge_mottu_2_semestre.challenge_mottu.repository.MotoRepository;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -12,6 +13,9 @@ import java.util.Optional;
 
 @Service
 public class GalpaoService {
+
+    @Autowired
+    private MotoRepository motoRepository;
 
     private final GalpaoRepository galpaoRepository;
 
@@ -47,9 +51,15 @@ public class GalpaoService {
     public boolean excluir(Long id) {
         Optional<Galpao> galpao = galpaoRepository.findById(id);
         if (galpao.isPresent()) {
-            galpaoRepository.delete(galpao.get());
             return true;
         }
+
+        boolean emUso = motoRepository.existsByGalpaoId(id);
+        if (emUso ) {
+            throw new IllegalStateException("Impossível excluir: Galpão está sendo usado em uma moto ou mais motos ");
+        }
+
+        galpaoRepository.delete(galpao.get());
         return false;
     }
 }
